# src/main/java/io/swagger/api/PaymentApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Payment;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-05-26T20:28:47.813Z")

@Validated
@Api(value = "payment", description = "the payment API")
@RequestMapping(value = "/v1")
public interface PaymentApi {

    @ApiOperation(value = "Make a payment order", nickname = "addPayment", notes = "", authorizations = {
        @Authorization(value = "payments_auth", scopes = {
            @AuthorizationScope(scope = "write:payments", description = "make a payment")
            })
    }, tags={ "payment", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "ressouce created and proceed") })
    @RequestMapping(value = "/payment",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Void> addPayment(@ApiParam(value = "Payment object" ,required=true )  @Valid @RequestBody Payment body);

}