# swaggerFinal.yml
swagger: "2.0"
info:
  description: " service payment "
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
- name: "PaymentOrder"
  description: "payment order class"
- name: "CreditCardInformation"
  description: "credit card which user uses to make payment "
- name: "user"
  description: "user who make payment or receive payment"
schemes:
- "https"
- "http"
paths:
  /payment:
    post:
      tags:
      - "payment"
      summary: "Make a payment order"
      description: ""
      operationId: "addPayment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Payment object"
        required: true
        schema:
          $ref: "#/definitions/Payment"
      responses:
        "201":
          description : "ressouce created and proceed"
       
      security:
      - payments_auth:
        - "write:payments"
securityDefinitions:
  payments_auth:
    type: "oauth2"
    authorizationUrl: "http://localhos.com/payments"
    flow: "implicit"
    scopes:
      write:payments: "make a payment"
definitions:
  PaymentOrder:
    type: "object"
    properties:
      sellerAccount:
        type: "string"
      amount:
        type: "number"
        format: "float64"
      currency:
        type: "string"
      paymentOrderId:
        type: "string"
  Payment:
    type: "object"
    required:
    - "buyer_info"
    - "checkout_id"
    - "creditCardInfo"
    - "paymentOrders"
    properties:
      buyer_info:
        $ref: "#/definitions/User"
      checkout_id:
        type: "string"
        example: "sdsfef1121"
      creditCardInfo:
        $ref: "#/definitions/CreditCardInformation"
      paymentOrders:
        type: "array"
        description: "list of payment"
        xml:
          name: "paymentOrder"
          wrapped: true
        items:
          $ref: "#/definitions/PaymentOrder"
    xml:
      name: "Payment"
  User:
    type: "object"
    required:
      - "id"
      - "firstName"
      - "lastName"
    properties:
      id:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
    xml:
      name: "User"
  CreditCardInformation:
    type: "object"
    required:
      - "id"
      - "cardNumber"
      - "cvv"
      - "expirationDate"
      - "ownerName"
    properties:
      id:
        type: "integer"
        format: "int64"
      cardNumber:
        type: "string"
      cvv: 
        type: "integer"
        format: "int32"
      expirationDate:
        type : "string"
        format : "date"
      ownerName : 
        type : "string"
    xml:
      name: "creditCardInformation"
  PaymentResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"